pyproject.toml
[tool.poetry]
name = "backendproject"
version = "0.1.0"
description = ""
authors = ["yoohannes <yohannesgidey420@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
django-filter = "^21.1"
black = "^22.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
